mentor fixes: transactional outbox, int money, domain invariants, seats dedup, docker healthchecks, phpunit

+++ app/src/Application/Dto/EventDetailsDto.php
@@ -17,8 +17,8 @@ readonly class EventDetailsDto
         public string $venueCity,
         public ?float $venueLatitude,
         public ?float $venueLongitude,
-        public float $priceMin,
-        public float $priceMax,
+        public int $priceMin,
+        public int $priceMax,
         public string $status,
         public array $availableSeats,
     ) {
